+++ semver-gradle-plugin/main/kotlin/internal/LastSemVer.kt
@@ -6,17 +6,14 @@ import org.gradle.api.Project
 
 internal val Project.lastSemVer: Version
     get() =
-        git.tagsInCurrentCommit(git.headCommit.commit.hash).lastResultVersion(tagPrefix, true)
-            ?: git.tagsInCurrentBranch.lastResultVersion(tagPrefix, false) ?: InitialVersion
+        git.tagsInCurrentCommit(git.headCommit.commit.hash).lastResultVersion(tagPrefix)
+            ?: git.tagsInCurrentBranch.lastResultVersion(tagPrefix) ?: InitialVersion
 
-internal fun List<GitRef.Tag>.lastResultVersion(
-    tagPrefix: String,
-    inCurrentCommit: Boolean
-): Version? =
+internal fun List<GitRef.Tag>.lastResultVersion(tagPrefix: String): Version? =
     asSequence()
         .filter { tag -> tag.name.startsWith(tagPrefix) }
         .map { tag -> tag.name.substringAfter(tagPrefix) }
         .map(Version.Companion::safe)
         .mapNotNull(Result<Version>::getOrNull)
         .toList()
-        .run { if (inCurrentCommit) maxOrNull() else lastOrNull() }
+        .run { maxOrNull() }